requirements.txt
appdirs==1.4.3
aspy.yaml==1.1.2
astroid==2.1.0
attrs==18.2.0
black==0.0
cfgv==1.4.0
Click==7.0
colorama==0.4.1
entrypoints==0.3
flake8==3.7.6
Flask==1.0.2
future==0.17.1
identify==1.2.2
importlib-metadata==0.8
importlib-resources==1.0.2
iso8601==0.1.12
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
mysql-connector==2.2.9
mysqlclient==1.4.2.post1
nodeenv==1.3.3
Pillow==6.*
pkg-resources==0.0.0
pre-commit==1.14.4
pycodestyle==2.5.0
pyfingerprint==0.1
pyflakes==2.1.0
pylint==2.2.2
pyserial==3.4
python-statemachine==0.7.1
PyYAML==3.13
RPi.GPIO==0.6.5
six==1.12.0
toml==0.10.0
typed-ast==1.3.1
virtualenv==16.4.1
Werkzeug==0.14.1
wrapt==1.11.1
zipp==0.3.3
